+++ src/FluxPwaGenerator.mjs
@@ -1,5 +1,7 @@
 /** @typedef {import("./Pwa/fileFilter.mjs").fileFilter} fileFilter */
+/** @typedef {import("./Pwa/getIconTemplateFile.mjs").getIconTemplateFile} getIconTemplateFile */
 /** @typedef {import("../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
+/** @typedef {import("../../flux-pwa-api/src/Pwa/Manifest.mjs").Manifest} Manifest */
 
 export class FluxPwaGenerator {
     /**
@@ -26,14 +28,33 @@ export class FluxPwaGenerator {
     }
 
     /**
-     * @param {string} fallback_icon_template_svg_file
+     * @param {string} web_root
+     * @param {fileFilter | null} filter_filter
+     * @param {boolean | null} ignore_jsdoc_files
+     * @returns {Promise<void>}
+     */
+    async deleteIgnoresFiles(web_root, filter_filter = null, ignore_jsdoc_files = null) {
+        await (await import("./Pwa/DeleteIgnoresFiles.mjs")).DeleteIgnoresFiles.new(
+            this
+        )
+            .deleteIgnoresFiles(
+                web_root,
+                filter_filter,
+                ignore_jsdoc_files
+            );
+    }
+
+    /**
+     * @param {getIconTemplateFile | string} get_icon_template_file
      * @param {string} manifest_json_file
      * @returns {Promise<void>}
      */
-    async generateIcons(fallback_icon_template_svg_file, manifest_json_file) {
-        await (await import("./Pwa/GenerateIcons.mjs")).GenerateIcons.new()
+    async generateIcons(get_icon_template_file, manifest_json_file) {
+        await (await import("./Pwa/GenerateIcons.mjs")).GenerateIcons.new(
+            this
+        )
             .generateIcons(
-                fallback_icon_template_svg_file,
+                get_icon_template_file,
                 manifest_json_file
             );
     }
@@ -47,6 +68,7 @@ export class FluxPwaGenerator {
      */
     async generateIndexHtmls(index_template_html_file, index_html_file, manifest_json_file, localization_folder = null) {
         await (await import("./Pwa/GenerateIndexHtmls.mjs")).GenerateIndexHtmls.new(
+            this,
             this.#flux_localization_api
         )
             .generateIndexHtmls(
@@ -65,6 +87,7 @@ export class FluxPwaGenerator {
      */
     async generateManifestJsons(manifest_template_json_file, manifest_json_file, localization_folder = null) {
         await (await import("./Pwa/GenerateManifestJsons.mjs")).GenerateManifestJsons.new(
+            this,
             this.#flux_localization_api
         )
             .generateManifestJsons(
@@ -85,13 +108,41 @@ export class FluxPwaGenerator {
      * @returns {Promise<void>}
      */
     async generateServiceWorker(service_worker_template_mjs_file, service_worker_mjs_file, web_root, application_cache_prefix, data = null, filter_filter = null, ignore_jsdoc_files = null) {
-        await (await import("./Pwa/GenerateServiceWorker.mjs")).GenerateServiceWorker.new()
+        await (await import("./Pwa/GenerateServiceWorker.mjs")).GenerateServiceWorker.new(
+            this
+        )
             .generateServiceWorker(
                 service_worker_template_mjs_file,
                 service_worker_mjs_file,
                 web_root,
                 application_cache_prefix,
                 data,
+                filter_filter,
+                ignore_jsdoc_files
+            );
+    }
+
+    /**
+     * @param {string} manifest_json_file
+     * @returns {Promise<Manifest>}
+     */
+    async getManifest(manifest_json_file) {
+        return (await import("./Pwa/GetManifest.mjs")).GetManifest.new()
+            .getManifest(
+                manifest_json_file
+            );
+    }
+
+    /**
+     * @param {string} web_root
+     * @param {fileFilter | null} filter_filter
+     * @param {boolean | null} ignore_jsdoc_files
+     * @returns {Promise<string[][]>}
+     */
+    async scanFiles(web_root, filter_filter = null, ignore_jsdoc_files = null) {
+        return (await import("./Pwa/ScanFiles.mjs")).ScanFiles.new()
+            .scanFiles(
+                web_root,
                 filter_filter,
                 ignore_jsdoc_files
             );