+++ src/Pwa/GenerateManifestJsons.mjs
@@ -1,28 +1,37 @@
-import { readFile, writeFile } from "node:fs/promises";
+import { writeFile } from "node:fs/promises";
 
 /** @typedef {import("../../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
+/** @typedef {import("../FluxPwaGenerator.mjs").FluxPwaGenerator} FluxPwaGenerator */
 
 export class GenerateManifestJsons {
     /**
      * @type {FluxLocalizationApi | null}
      */
     #flux_localization_api;
+    /**
+     * @type {FluxPwaGenerator}
+     */
+    #flux_pwa_generator;
 
     /**
+     * @param {FluxPwaGenerator} flux_pwa_generator
      * @param {FluxLocalizationApi | null} flux_localization_api
      * @returns {GenerateManifestJsons}
      */
-    static new(flux_localization_api = null) {
+    static new(flux_pwa_generator, flux_localization_api = null) {
         return new this(
+            flux_pwa_generator,
             flux_localization_api
         );
     }
 
     /**
+     * @param {FluxPwaGenerator} flux_pwa_generator
      * @param {FluxLocalizationApi | null} flux_localization_api
      * @private
      */
-    constructor(flux_localization_api) {
+    constructor(flux_pwa_generator, flux_localization_api) {
+        this.#flux_pwa_generator = flux_pwa_generator;
         this.#flux_localization_api = flux_localization_api;
     }
 
@@ -43,7 +52,9 @@ export class GenerateManifestJsons {
             );
         }
 
-        const manifest = JSON.parse(await readFile(manifest_template_json_file, "utf8"));
+        const manifest = await this.#flux_pwa_generator.getManifest(
+            manifest_template_json_file
+        );
 
         const manifest_json_file_dot_pos = manifest_json_file.lastIndexOf(".");